+++ kernel/string.c
@@ -5,7 +5,7 @@
 
 __nonnull(1) void *memset(void *dst, int c, uint32_t n)
 {
-	if ((int)dst % 4 == 0 && n % 4 == 0) {
+	if ((size_t)dst % 4 == 0 && n % 4 == 0) {
 		c &= 0xFF;
 		stosl(dst, (c << 24) | (c << 16) | (c << 8) | c, n / 4);
 	} else
